Move some design picker logic into selector
The logic which gets the billing product for a given marketplace theme
has been moved into a central selector so that it can be used in the
themeshowcase too.
